Feature #21526 Teacher: Create Forum  -Created new "Add Forum" page.  -Added date picker,time picker,radio list,check boxes,labels on "Add Forum" page.  -Written methods on models to store new forum data in daatabase on clicks of "Create Forum" button.  -Implemented a logic to display created forum.  Teacher: Delete Forum  -Shown pop up window when user clicks on delete forum link.  -Written methods on models to delete forum data from database on clicks of "Confirm" button.

+++ models/Thread.php
@@ -45,6 +45,16 @@ class Thread extends BaseImasForumThreads
     public static function deleteThreadById($id)
     {
         $thread = Thread::findOne(['id' => $id]);
-        $thread->delete();
+        if($thread){
+            $thread->delete();
+        }
+
     }
+    public static function moveAndUpdateThread($forumId,$threadId)
+    {
+        $ForumPost = Thread::findOne(['id' => $threadId]);
+        $ForumPost->forumid = $forumId;
+        $ForumPost->save();
+    }
+
 }
